Board 클래스 리팩토링 및 테스트 코드 수정 Board 클래스에서 rankList getter 생성 Board 클래스의 기물 개수, 점수, 현재 보드 상황 출력, 리스트 정렬 등의 기능들을 따로 관리하는 BoardStatus 클래스 생성하여 분리 그에 맞춰 테스트 코드와 main 메서드 수정함.
